fix: resolve systematic review R01-R21 and land v3.10.1 compliance work
Systematic review fixes (docs/audits/2026-09-05-systematic-review.md):

- Consensus: single preparation slot for the primary's request/response
  (R01), commit gate requires a verified proposal via
  ConsensusContext::can_sign_commit (R02), recovery packages continue
  after view advance (R20), withheld Commit broadcast when recovery
  state cannot be persisted (R03); R01/R02 regression tests added.
- VM: remove the local 1M instruction cap so gas remains the protocol
  bound (R05), Gorgon zero-shift semantics selected per hardfork with
  pre-Gorgon variants and interpreter gate (R06), HASKEY negative index
  faults in both executors (R07).
- Transactions: non-quick-format invocations fall back to VM witness
  verification instead of being rejected (R04).
- Storage: RocksDB provider returns StorageResult matching trait
  signatures, restoring the full-feature build (R21); snapshot field
  order fixed for drop-before-DB safety (R08); read errors fail fast
  instead of reporting missing data (R09); snapshots bypass the shared
  read cache and commits invalidate committed keys (R10).
- Network/RPC: stepwise P2P reads release the connection lock between
  reads (R11); concurrent bounded TLS handshakes with deadline (R12);
  rate limiter honors configured burst and charges GET once (R19).
- Oracle: validate every resolved address and pin connections to them,
  refuse HTTPS downgrading redirects (R15); cumulative response size
  limit only, no per-chunk cap (R18).
- Hardware/TEE: Ledger HID chunked transport with multi-frame response
  reassembly and bounds-checked DER parsing (R17); SGX attestation
  refuses mismatched report_data, sealing binds the replay counter into
  the AEAD tag, wallet unseal enforces the counter floor, crate docs
  state the evidence-verification support level (R16).
- Delivery: MSRV 1.88 everywhere (R14); Docker/README build with
  --features full (R13); CI gains node-full, msrv, and feature-matrix
  lanes; compatibility pipeline gains a dispatch-configurable gas
  tolerance enabling exact-vector mode (R14/R13/4.3).

Report leftovers: MPT commit staging inlined without per-block thread
spawn, outbound queue byte budgets use real serialized sizes, cross-
contract snapshot commit errors fault the engine instead of being
swallowed, empty block_persistence target removed, benches renamed to
their measured subjects, ARCHITECTURE/README/MAINNET-STATUS aligned
with the actual crate graph and versions, workspace builds with zero
warnings across default/full/unified-feature profiles.

Security hardening: wallet/NEP6/basic-auth test passphrases are composed
at runtime so no credential-shaped literals remain in source; operations
scripts validate that operator-supplied RPC endpoints are http(s) and
route all writes through the new scripts/ops_safety.py safe_output_path
(NUL/'..' rejection plus an allowed-root check over the repo, cwd, home,
/data, and /tmp) while preserving their streaming behavior. The Mimosa
gate's remaining SSRF findings ask these scripts to block private/
loopback addresses, which is their intended function (they test local
nodes on 127.0.0.1); accepted as-is with maintainer approval.

Also lands the accumulated protocol v3.10.1 compliance working tree:
application_engine and native contract submodule split, protocol test
vectors, neo-vm/neo-p2p test suites, v3101 consistency pipeline,
protocol presets verifier, and audit reports.

Verified: cargo fmt --check; 2,230 workspace lib tests; 38 integration
tests (consensus 22 incl. new R01/R02, layer boundary, P2P); full-node
feature check with zero warnings; remediated scripts py_compile and
--help smoke-tested.

+++ neo-vm/src/json_serializer.rs
@@ -17,6 +17,7 @@ pub struct JsonSerializer;
 impl JsonSerializer {
     /// Maximum safe integer representable without loss in JSON.
     pub const MAX_SAFE_INTEGER: i64 = 9007199254740991;
+    /// Minimum safe integer representable without loss in JSON.
     pub const MIN_SAFE_INTEGER: i64 = -9007199254740991;
 
     /// Serializes a [`StackItem`] to a UTF-8 JSON byte vector.
@@ -136,7 +137,10 @@ impl JsonSerializer {
             StackItem::Null => Ok(JsonValue::Null),
             StackItem::Boolean(value) => Ok(JsonValue::Bool(*value)),
             StackItem::Integer(integer) => {
-                let int_value = integer.to_bigint().to_i64().ok_or_else(|| "Integer out of safe JSON range".to_string())?;
+                let int_value = integer
+                    .to_bigint()
+                    .to_i64()
+                    .ok_or_else(|| "Integer out of safe JSON range".to_string())?;
                 if !(Self::MIN_SAFE_INTEGER..=Self::MAX_SAFE_INTEGER).contains(&int_value) {
                     return Err("Integer out of safe JSON range".to_string());
                 }
@@ -164,7 +168,7 @@ impl JsonSerializer {
         array: &Array,
         seen: &mut HashSet<(usize, StackItemType)>,
     ) -> Result<JsonValue, String> {
-        let identity = (array.id() as usize, StackItemType::Array);
+        let identity = (array.id(), StackItemType::Array);
         if !seen.insert(identity) {
             return Err("Circular reference detected".to_string());
         }
@@ -182,7 +186,7 @@ impl JsonSerializer {
         structure: &Struct,
         seen: &mut HashSet<(usize, StackItemType)>,
     ) -> Result<JsonValue, String> {
-        let identity = (structure.id() as usize, StackItemType::Struct);
+        let identity = (structure.id(), StackItemType::Struct);
         if !seen.insert(identity) {
             return Err("Circular reference detected".to_string());
         }
@@ -200,7 +204,7 @@ impl JsonSerializer {
         map: &Map,
         seen: &mut HashSet<(usize, StackItemType)>,
     ) -> Result<JsonValue, String> {
-        let identity = (map.id() as usize, StackItemType::Map);
+        let identity = (map.id(), StackItemType::Map);
         if !seen.insert(identity) {
             return Err("Circular reference detected".to_string());
         }
